universa v0.1.0 Universa.Channel

Types

channel()

channel() :: String.t() | atom()

entities()

entities() :: [String.t()] | String.t()

entity()

entity() :: String.t()

Functions

add(channel, entities)

add(channel(), entities()) :: :ok

Add entity to channel

clear(channel)

clear(channel()) :: :ok

Remove all entities from this channel

memberof(entity)

memberof(entity()) :: [map()]

Retrieve all channels this entity is a member of

members(channel)

members(channel()) :: [map()]

Retrieve all entities that are a member of this channel

remove(channel, entity)

remove(channel(), entity()) :: :ok

Remove entity from channel