universa v0.1.0 Universa.Database.Account

Automatically generated database table.

This is an automatically generated file, sadly we can’t provide detailed documentation for this module at this time. This file was generated so the code knows how to create and use the associated table, using “Elixir.Universa.Database.Account” as the name of the table.

Please see Universa.Database.Table for more information.

Functions

defaults()

defaults() :: list()

Returns a list with all keys and their defaults.

Example

iex> Universa.Database.Entity.defaults
[uuid: "", systems: []]

keys()

keys() :: list()

Returns all keys for this table.

Example

iex> Universa.Database.Entity.keys
[:uuid, :systems]

primary_key()

primary_key() :: atom()

Returns the primary key for this table, this key must be unique for every entry.

Example

iex> Universa.Database.Entity.primary_key
:uuid